pyproject.toml
[tool.poetry]
name = "pseudo_opendns"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dnslib = "^0.9.19"
python-dotenv = "^0.20.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
